% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leukemia.test.R
\docType{data}
\name{leukemia.test}
\alias{leukemia.test}
\title{Gene expression Leukemia test data set from Golub et al. (1999)}
\format{
## `leukemia.test`
A data frame with 34 rows and 7130 columns:
\describe{
  ... 
  \item{V7130}{class label}
}
}
\source{
<http://wwwprod.broadinstitute.org/cgi-bin/cancer/datasets.cgi>
}
\usage{
leukemia.test
}
\description{
Gene expression test data of 7129 genes from 34 patients 
with acute leukemias (20 in class Acute Lymphoblastic Leukemia and 14 in class Acute Myeloid Leukemia) from the microarray study of Golub et al. (1999).
}
\keyword{datasets}
