% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leukemia.train.R
\docType{data}
\name{leukemia.train}
\alias{leukemia.train}
\title{Gene expression Leukemia training data set from Golub et al. (1999)}
\format{
## `leukemia.train`
A data frame with 38 rows and 7130 columns:
\describe{
  ... 
  \item{V7130}{class label}
}
}
\source{
<http://wwwprod.broadinstitute.org/cgi-bin/cancer/datasets.cgi>
}
\usage{
leukemia.train
}
\description{
Gene expression training data of 7129 genes from 38 patients 
with acute leukemias (27 in class Acute Lymphoblastic Leukemia and 11 in class Acute Myeloid Leukemia) from the microarray study of Golub et al. (1999).
}
\keyword{datasets}
