% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostate.test.R
\docType{data}
\name{prostate.test}
\alias{prostate.test}
\title{Gene expression Prostate Cancer test data set from Singh et al. (2002)}
\format{
## `prostate.test`
A data frame with 34 rows and 12601 columns:
\describe{
  ... 
  \item{V12601}{class label}
}
}
\source{
<http://wwwprod.broadinstitute.org/cgi-bin/cancer/datasets.cgi>
}
\usage{
prostate.test
}
\description{
Gene expression training data of 12600 genes from 25 patients with prostate tumors and 9 normal specimens from the microarray study of Singh et al. (2002).
}
\keyword{datasets}
