% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostate.train.R
\docType{data}
\name{prostate.train}
\alias{prostate.train}
\title{Gene expression Prostate Cancer training data set from Singh et al. (2002)}
\format{
## `prostate.train`
A data frame with 102 rows and 12601 columns:
\describe{
  ... 
  \item{V12601}{class label}
}
}
\source{
<http://wwwprod.broadinstitute.org/cgi-bin/cancer/datasets.cgi>
}
\usage{
prostate.train
}
\description{
Gene expression training data of 12600 genes from 52 patients with prostate tumors and 50 normal specimens from the microarray study of Singh et al. (2002).
}
\keyword{datasets}
