\name{leukemia.test}
\alias{leukemia.test}
\title{Gene expression Leukemia testing dataset from Golub et al. (1999)}
\description{Gene expression testing data of 7129 genes from 34 patients 
with acute leukemias from the microarray study of Golub et al. (1999).
} 
\usage{data(leukemia.test)}
\format{A data frame with 34 observations on 7129 variables.}
\source{http://wwwprod.broadinstitute.org/cgi-bin/cancer/datasets.cgi}
\references{
Golub et al. (1999) Molecular Classification of Cancer: Class Discovery and 
Class Prediction by Gene Expression Monitoring. \emph{Science}, \bold{286}, 
531-537.
} 
\keyword{datasets}
