\name{leukemia.train}
\alias{leukemia.train}
\title{Gene expression Leukemia training data set from Golub et al. (1999)}
\description{Gene expression training data of 7129 genes from 38 patients 
with acute leukemias (27 in class Acute Lymphoblastic Leukemia and 11 in 
class Acute Myeloid Leukemia) from the microarray study of Golub et al. (1999).
} 
\usage{data(leukemia.train)}
\format{A data frame with 38 observations on 7129 variables.}
\source{http://wwwprod.broadinstitute.org/cgi-bin/cancer/datasets.cgi}
\references{
Golub et al. (1999) Molecular Classification of Cancer: Class Discovery and 
Class Prediction by Gene Expression Monitoring. \emph{Science}, \bold{286}, 
531-537.
} 
\keyword{datasets}
