\name{prostate.test}
\alias{prostate.test}
\title{Gene expression Prostate Cancer testing data set from Singh et al. (2002)}
\description{Gene expression testing data of 12600 genes from 25 patients with 
prostate tumors and 9 normal specimens from the microarray study of 
Singh et al. (2002).
} 
\usage{data(prostate.test)}
\format{A data frame with 34 observations on 12600 variables.}
\source{http://wwwprod.broadinstitute.org/cgi-bin/cancer/datasets.cgi}
\references{
Singh et al. (2002) Gene Expression Correlates of Clinical Prostate Cancer 
Behavior. \emph{Cancer Cell}, \bold{1}, 203-209.
} 
\keyword{datasets}


