\name{prostate.train}
\alias{prostate.train}
\title{Gene expression Prostate Cancer training data set from Singh et al. (2002)}
\description{Gene expression training data of 12600 genes from 52 patients with 
prostate tumors and 50 normal specimens from the microarray study of 
Singh et al. (2002).
} 
\usage{data(prostate.train)}
\format{A data frame with 102 observations on 12600 variables.}
\source{http://wwwprod.broadinstitute.org/cgi-bin/cancer/datasets.cgi}
\references{
Singh et al. (2002) Gene Expression Correlates of Clinical Prostate Cancer 
Behavior. \emph{Cancer Cell}, \bold{1}, 203-209.
} 
\keyword{datasets}
